+++ src/constants/shared.ts
@@ -1,27 +1,5 @@
 import JSBI from 'jsbi'
-
-export type BigintIsh = JSBI | bigint | string
-
-export enum ChainId {
-  MAINNET = 1,
-  ROPSTEN = 3,
-  RINKEBY = 4,
-  GOERLI = 5,
-  KOVAN = 42,
-  POLYGON = 137,
-  POLYGON_MUMBAI = 80001,
-}
-
-export enum TradeType {
-  EXACT_INPUT,
-  EXACT_OUTPUT,
-}
-
-export enum Rounding {
-  ROUND_DOWN,
-  ROUND_HALF_UP,
-  ROUND_UP,
-}
+import { SolidityType } from 'types'
 
 export const FACTORY_ADDRESS = '0x69bd16aE6F507bd3Fc9eCC984d50b04F029EF677'
 
@@ -38,11 +16,6 @@ export const TEN = JSBI.BigInt(10)
 export const _100 = JSBI.BigInt(100)
 export const _997 = JSBI.BigInt(997)
 export const _1000 = JSBI.BigInt(1000)
-
-export enum SolidityType {
-  uint8 = 'uint8',
-  uint256 = 'uint256',
-}
 
 export const SOLIDITY_TYPE_MAXIMA = {
   [SolidityType.uint8]: JSBI.BigInt('0xff'),
