+++ src/entities/fractions/fraction.ts
@@ -3,9 +3,9 @@ import JSBI from 'jsbi'
 import _Decimal from 'decimal.js-light'
 import _Big, { RoundingMode } from 'big.js'
 import toFormat from 'toformat'
-import { BigintIsh, Rounding } from 'constants'
 import { ONE } from 'constants'
 import { parseBigintIsh } from 'utils'
+import { BigintIsh, Rounding } from 'types'
 
 const Decimal = toFormat(_Decimal)
 const Big = toFormat(_Big)
