+++ src/entities/nativeCurrency.ts
@@ -1,6 +1,7 @@
-import { ChainId, WRAPPED_NATIVE_CURRENCY } from 'constants'
+import { WRAPPED_NATIVE_CURRENCY } from 'constants'
 import { BaseCurrency, Token, Currency } from 'entities'
 import invariant from 'tiny-invariant'
+import { ChainId } from 'types'
 
 export class NativeCurrency extends BaseCurrency {
   public readonly isNative: true = true
