+++ src/entities/pair.ts
@@ -3,21 +3,11 @@ import invariant from 'tiny-invariant'
 import JSBI from 'jsbi'
 import { pack, keccak256 } from '@ethersproject/solidity'
 import { getCreate2Address } from '@ethersproject/address'
-import {
-  BigintIsh,
-  FACTORY_ADDRESS,
-  INIT_CODE_HASH,
-  MINIMUM_LIQUIDITY,
-  ZERO,
-  ONE,
-  FIVE,
-  _997,
-  _1000,
-  ChainId,
-} from 'constants'
+import { FACTORY_ADDRESS, INIT_CODE_HASH, MINIMUM_LIQUIDITY, ZERO, ONE, FIVE, _997, _1000 } from 'constants'
 import { sqrt, parseBigintIsh } from 'utils'
-import { InsufficientReservesError, InsufficientInputAmountError } from 'errors'
+import { InsufficientReservesError, InsufficientInputAmountError } from 'types/errors'
 import { Token } from 'entities'
+import { BigintIsh, ChainId } from 'types'
 
 let PAIR_ADDRESS_CACHE: { [token0Address: string]: { [token1Address: string]: string } } = {}
