+++ src/index.ts
@@ -2,10 +2,6 @@ import JSBI from 'jsbi'
 export { JSBI }
 
 export {
-  BigintIsh,
-  ChainId,
-  TradeType,
-  Rounding,
   FACTORY_ADDRESS,
   INIT_CODE_HASH,
   MINIMUM_LIQUIDITY,
@@ -13,7 +9,7 @@ export {
   WRAPPED_NATIVE_CURRENCY,
   WETH,
 } from 'constants'
-export * from 'errors'
+export * from 'types'
 export * from 'entities'
 export * from 'router'
 export * from 'fetcher'
