+++ src/types/errors.ts
