+++ src/types/index.ts
@@ -1,0 +1,2 @@
+export * from 'types/errors'
+export * from 'types/shared'
