+++ src/types/shared.ts
@@ -1,0 +1,29 @@
+import JSBI from 'jsbi'
+
+export type BigintIsh = JSBI | bigint | string
+
+export enum ChainId {
+  MAINNET = 1,
+  ROPSTEN = 3,
+  RINKEBY = 4,
+  GOERLI = 5,
+  KOVAN = 42,
+  POLYGON = 137,
+  POLYGON_MUMBAI = 80001,
+}
+
+export enum TradeType {
+  EXACT_INPUT,
+  EXACT_OUTPUT,
+}
+
+export enum Rounding {
+  ROUND_DOWN,
+  ROUND_HALF_UP,
+  ROUND_UP,
+}
+
+export enum SolidityType {
+  uint8 = 'uint8',
+  uint256 = 'uint256',
+}
