+++ src/utils.ts
@@ -2,7 +2,8 @@ import invariant from 'tiny-invariant'
 import warning from 'tiny-warning'
 import JSBI from 'jsbi'
 import { getAddress } from '@ethersproject/address'
-import { BigintIsh, ZERO, ONE, TWO, THREE, SolidityType, SOLIDITY_TYPE_MAXIMA } from 'constants'
+import { ZERO, ONE, TWO, THREE, SOLIDITY_TYPE_MAXIMA } from 'constants'
+import { BigintIsh, SolidityType } from 'types'
 
 export function validateSolidityTypeInstance(value: JSBI, solidityType: SolidityType): void {
   invariant(JSBI.greaterThanOrEqual(value, ZERO), `${value} is not a ${solidityType}.`)
